# api/api.yaml
openapi: 3.1.0
info:
  title: EST Back API
  description: Все запросы от прокси к бэку
  version: 1.0.0
servers:
  - url: 'https://pursit.ru'
security:
  - jwt: [ ]
paths:
...
  /backend/sitter:
    get:
      tags:
        - sitters
      summary: Получение ситтера по id
      operationId: getSittersByIds
      parameters:
        - name: sittersIds
          in: query
          required: true
          schema:
            type: array
            items:
              type: string
              format: uuid
      responses:
        200:
          description: Ситтер
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SitterDto"
...
  /backend/sitter/request:
    post:
      tags:
        - sitters
      summary: Отправка заявки на передержку
      operationId: sendRequest
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SitterRequestDto"
      responses:
        200:
          description: Заявка отправлена
components:
  schemas:
...
    SitterDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          $ref: "#/components/schemas/SitterType"
        name:
          description: Имя ситтера
          type: string
        avatarUrl:
          description: Ссылка на аватар
          type: string
        shortDescription:
          description: Краткое описание
          type: string
        longDescription:
          description: Полное описание
          type: string
        price:
          description: Цена за сутки
          type: number
        experience:
          description: Стаж ситтера
          type: number
        photosUrls:
          description: Ссылки на фотографии
          type: array
          items:
            type: string
        readyComeItself:
          description: Готов сам приехать
          type: boolean
        takesCats:
          type: boolean
        takesDogs:
          type: boolean
        geolocation:
          $ref: "#/components/schemas/GeolocationDto"
    SitterRequestDto:
      type: object
      properties:
        sitterIds:
          description: id выбранных ситтеров
          type: array
          items:
            type: string
            format: uuid
        fullName:
          description: Имя владельца
          type: string
        phone:
          description: Номер телефона клиента
          type: string
        email:
          description: Почта клиента
          type: string
        tg:
          description: Ник в телеграме клиента
          type: string
        reportsTarget:
          $ref: "#/components/schemas/ReportsTarget"
        startDate:
          description: Дата начала передержки
          type: string
          format: date-time
        endDate:
          description: Дата окончания передержки
          type: string
          format: date-time
        petName:
          description: Имя питомца
          type: string
        isCat:
          description: Является ли питомец кошкой, false - собака
          type: boolean
        breed:
          description: Порода питомца
          type: string
        isMale:
          description: Является ли питомец мальчиком, false - девочкой
          type: boolean
        height:
          description: Высота питомца в холке
          type: string
        hasVaccinations:
          description: Наличие прививок у питомца
          type: boolean
        otherMessage:
          description: Другие пожелания клиента
          type: string
        needTake:
          description: Вашего питомца забрать или вы привезете его самостоятельно?
          type: boolean
        addressTake:
          description: Укажите адрес откуда забрать питомца
          type: string
    ReportsTarget:
      type: string
      enum:
        - tg
        - whatsapp
        - email
    GeolocationDto:
      type: object
      properties:
        latitude:
          description: Широта
          type: number
        longitude:
          description: Долгота
          type: number
        name:
          description: Название
          type: string
    SitterType:
      type: string
      enum:
        - moral-pet-sitting
        - pet-sitting
        - pet-health
        - pet-moving